+++ lib/kb_gtdbtk/core/gtdbtk_runner.py
@@ -98,8 +98,13 @@ def _process_output_files(temp_output, out_dir, id_to_name):
             for item in sj['data']:
                 # no blank fields.  messes up datatables in index.html
                 for key in item.keys():
+                    print ("RESULTS: k: '"+key+"' val: '"+str(item[key])+"'")  # DEBUG
                     if not item.get(key):
                         item[key] = '-'  # note: this resets data in sj
+
+                # sometimes missing aa_percent field, which disrupts datatables report
+                if 'aa_percent' not in item:
+                    item['aa_percent'] = '-'
 
                 # field 'Name' was changed to 'name'
                 #key = 'Name' if 'Name' in item else 'user_genome'